+++ types.go
@@ -5,6 +5,7 @@ import "errors"
 var (
 	ErrNotFound      = errors.New("option not found")
 	ErrTypeMismatch  = errors.New("wrong type expected from option")
+	ErrNilValue      = errors.New("nil value")
 	ErrDuplicatedKey = errors.New("option has already been set")
 )